alterei lat e long para Float

+++ app/models/sensor.rb
@@ -6,8 +6,8 @@ class Sensor
   field :description, type: String
   field :min, type: Integer, default: 0
   field :max, type: Integer, default: 100
-  field :latitude, type: Integer, default: 0
-  field :longitude, type: Integer, default: 0
+  field :latitude, type: Float, default: 0
+  field :longitude, type: Float, default: 0
 
   embedded_in :zones
